一部エンドポイントの外部ファイル化をしました
